requirements.txt
PyQt5>=5.15.0
qgis>=3.22.0
pytest>=7.0.0
pytest-qt>=4.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0
coverage>=6.0.0
mock>=4.0.0
python-dotenv>=1.0.0 